# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api-json":{"get":{"operationId":"AppController_code","summary":"获取接口文档json数据","parameters":[],"responses":{"200":{"description":""}}}},"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/user/init":{"get":{"operationId":"UserController_initData","summary":"初始化用户、角色、权限","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/code":{"get":{"operationId":"UserController_code","summary":"获取图文验证码","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/emailCode":{"get":{"operationId":"UserController_emailCode","summary":"获取邮箱验证码","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/login":{"post":{"operationId":"UserController_login","summary":"用户登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["用户"]}},"/api/user/refresh":{"get":{"operationId":"UserController_refresh","summary":"刷新token","parameters":[{"name":"refresh_token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/register":{"post":{"operationId":"UserController_create","summary":"创建用户","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["用户"]}},"/api/user/{id}":{"delete":{"operationId":"UserController_delete","summary":"删除用户","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["用户"]},"patch":{"operationId":"UserController_update","summary":"修改用户","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/info/{id}":{"get":{"operationId":"UserController_getUserInfo","summary":"获取用户信息","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/findAll":{"post":{"operationId":"UserController_findAll","summary":"查找所有用户带分页带keyword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/findAllUserDto"}}}},"responses":{"201":{"description":""}},"tags":["用户"]}},"/api/user/addFriend":{"post":{"operationId":"UserController_addFriend","summary":"添加好友","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddFriendDto"}}}},"responses":{"201":{"description":""}},"tags":["用户"]}},"/api/user/friendList":{"get":{"operationId":"UserController_getFriendList","summary":"获取好友列表","parameters":[{"name":"type","required":false,"in":"query","description":"notice-待确认和已拒绝的列表, friend-已添加的好友列表, black-黑名单列表, all或不传-返回所有类型","schema":{"enum":["notice","friend","black","all"],"type":"string"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/friend/{friendId}":{"patch":{"operationId":"UserController_updateFriend","summary":"更新好友信息","parameters":[{"name":"friendId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFriendDto"}}}},"responses":{"200":{"description":""}},"tags":["用户"]},"delete":{"operationId":"UserController_deleteFriend","summary":"删除好友","parameters":[{"name":"friendId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/user/chat/{friendId}":{"post":{"operationId":"UserController_createChatList","summary":"创建聊天关系","parameters":[{"name":"friendId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["用户"]}},"/api/user/chat/list":{"get":{"operationId":"UserController_getChatList","summary":"获取聊天列表","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"]}},"/api/menu":{"post":{"operationId":"MenuController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMenuDto"}}}},"responses":{"201":{"description":""}},"tags":["菜单"]},"get":{"operationId":"MenuController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["菜单"]}},"/api/menu/{id}":{"get":{"operationId":"MenuController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["菜单"]},"patch":{"operationId":"MenuController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMenuDto"}}}},"responses":{"200":{"description":""}},"tags":["菜单"]},"delete":{"operationId":"MenuController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["菜单"]}},"/api/upload":{"post":{"operationId":"UploadController_upload","summary":"单文件上传","parameters":[],"responses":{"201":{"description":""}},"tags":["上传"]},"get":{"operationId":"UploadController_getAllFile","summary":"查询所有文件","parameters":[],"responses":{"200":{"description":""}},"tags":["上传"]}},"/api/upload/uploadFile":{"post":{"operationId":"UploadController_uploadFile","summary":"分片上传","parameters":[],"responses":{"201":{"description":""}},"tags":["上传"]}},"/api/upload/merge":{"get":{"operationId":"UploadController_merge","summary":"合并文件","parameters":[{"name":"name","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["上传"]}},"/api/upload/addFile":{"post":{"operationId":"UploadController_create","summary":"将上传的文件的对应关系存储到数据库","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFileDto"}}}},"responses":{"201":{"description":""}},"tags":["上传"]}},"/api/upload/{id}":{"get":{"operationId":"UploadController_getFile","summary":"根据id查询文件","parameters":[],"responses":{"200":{"description":""}},"tags":["上传"]},"delete":{"operationId":"UploadController_deleteFile","summary":"删除文件","parameters":[],"responses":{"200":{"description":""}},"tags":["上传"]}},"/api/upload/download/{id}":{"get":{"operationId":"UploadController_export","summary":"下载文件","parameters":[],"responses":{"200":{"description":""}},"tags":["上传"]}},"/message":{"post":{"operationId":"MessageController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMessageDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"MessageController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/message/{id}":{"get":{"operationId":"MessageController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"MessageController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMessageDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"MessageController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Cats example","description":"The cats API description","version":"1.0","contact":{}},"tags":[{"name":"cats","description":""}],"servers":[],"components":{"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string","description":"用户名","example":"admin"},"email":{"type":"string","description":"邮箱","example":"email@qq.com"},"password":{"type":"string","description":"密码","example":"123456"},"code":{"type":"string","description":"验证码"}},"required":["username","email","password","code"]},"CreateUserDto":{"type":"object","properties":{"id":{"type":"number","description":"用户id","example":1},"username":{"type":"string","description":"用户名(系统自动生成)","example":"1234567890"},"nickname":{"type":"string","description":"昵称","example":"张三"},"password":{"type":"string","description":"密码","example":"123456"},"email":{"type":"string","description":"邮箱","example":"1640551913@qq.com"},"role":{"type":"number","description":"角色","example":"1"},"phone":{"type":"string","description":"手机号","example":"13888888888"},"state":{"type":"number","description":"状态","example":1},"avatar":{"type":"string","description":"头像"},"online":{"type":"string","description":"在线状态","example":"0"},"gender":{"type":"string","description":"性别","example":"0","enum":["0","1"],"default":"0"},"desc":{"type":"string","description":"描述","example":"这是一段个人描述"},"address":{"type":"string","description":"地址","example":"北京市朝阳区"}},"required":["id","username","nickname","password","email","role","phone","state","avatar","online","gender","desc","address"]},"UpdateUserDto":{"type":"object","properties":{"id":{"type":"number","description":"用户id","example":1},"username":{"type":"string","description":"用户名(系统自动生成)","example":"1234567890"},"nickname":{"type":"string","description":"昵称","example":"张三"},"password":{"type":"string","description":"密码","example":"123456"},"email":{"type":"string","description":"邮箱","example":"1640551913@qq.com"},"role":{"type":"number","description":"角色","example":"1"},"phone":{"type":"string","description":"手机号","example":"13888888888"},"state":{"type":"number","description":"状态","example":1},"avatar":{"type":"string","description":"头像"},"online":{"type":"string","description":"在线状态","example":"0"},"gender":{"type":"string","description":"性别","example":"0","enum":["0","1"],"default":"0"},"desc":{"type":"string","description":"描述","example":"这是一段个人描述"},"address":{"type":"string","description":"地址","example":"北京市朝阳区"}}},"findAllUserDto":{"type":"object","properties":{"page":{"type":"number","description":"页码","example":1},"pageSize":{"type":"number","description":"每页条数","example":10},"keyWord":{"type":"string","description":"关键字"}},"required":["page","pageSize","keyWord"]},"AddFriendDto":{"type":"object","properties":{"friendId":{"type":"number","description":"好友ID"},"remark":{"type":"string","description":"好友备注"},"friendRemark":{"type":"string","description":"对方的备注"},"desc":{"type":"string","description":"描述"}},"required":["friendId"]},"UpdateFriendDto":{"type":"object","properties":{"remark":{"type":"string","description":"好友备注"},"status":{"type":"string","description":"状态"}},"required":["remark","status"]},"CreateMenuDto":{"type":"object","properties":{"menuId":{"type":"number","description":"菜单id","example":1},"parentMenuId":{"type":"number","description":"父菜单id","example":1},"title":{"type":"string","description":"菜单名"},"path":{"type":"string","description":"URL路径"},"name":{"type":"string","description":"标题"},"component":{"type":"string","description":"组件路径"},"redirect":{"type":"string","description":"重定向"},"icon":{"type":"string","description":"图标"},"hidden":{"type":"boolean","description":"是否隐藏"}},"required":["menuId","parentMenuId","title","path","name","component","redirect","icon","hidden"]},"UpdateMenuDto":{"type":"object","properties":{"menuId":{"type":"number","description":"菜单id","example":1},"parentMenuId":{"type":"number","description":"父菜单id","example":1},"title":{"type":"string","description":"菜单名"},"path":{"type":"string","description":"URL路径"},"name":{"type":"string","description":"标题"},"component":{"type":"string","description":"组件路径"},"redirect":{"type":"string","description":"重定向"},"icon":{"type":"string","description":"图标"},"hidden":{"type":"boolean","description":"是否隐藏"}}},"CreateFileDto":{"type":"object","properties":{"id":{"type":"number","description":"文件id","example":1},"fileName":{"type":"string","description":"文件名"},"desc":{"type":"string","description":"描述"}},"required":["id","fileName","desc"]},"CreateMessageDto":{"type":"object","properties":{}},"UpdateMessageDto":{"type":"object","properties":{}}}}}